Table.model & rename 'tabe' directive into table
