+++ src/app/components/table/table.directive.ts
@@ -1,39 +1,6 @@
 import {IAugmentedJQuery, IDirective, IDirectiveFactory, ILocationService, IAttributes} from 'angular';
 import {TableCtrl, ITableScope} from './table.controller';
 
-export interface IRowAffairs  {
-  reason: string;
-}
-
-export interface IRowUsed {
-  time: Date;
-  patient: string;
-}
-
-export interface IRowFree {
-  time: Date;
-}
-
-export interface IColumn {
-  date: Date;
-  doctor: string;
-  specialty: string;
-  address: string;
-}
-
-export interface IColumnBusy extends IColumn {
-  busy: string;
-}
-
-export interface IColumnFree extends IColumn {
-  interval: string;
-  rows: Row[];
-}
-
-export type Column = IColumnBusy | IColumnFree;
-export type Row = IRowAffairs | IRowFree | IRowUsed;
-
-
 export class tableDirective implements IDirective {
   public restrict = 'E';
   public template = require('./table.html');
